develop_parallel: C++ parts of CUPyDO compiled on windows. .gitignore: *pyd added. CMakeLists.txt: installation instructions added. Now everything generated by the compiler is in the 'build' folder and the installer manages the copy of 'CCupydo.py' and '_CCupydo.pyd' into the 'CUPyDO/ccupydo' folder. FSICoupler.py: 'return value' in mpiAllReduce replaces 'return sendBuff' which was wrong for serial use. include/CCupydo.i: small changes to enhance robustness (i.e. disown() feature for directors which avoids memory leaks). cManager.h/cpp: correction of a bug, 'double' passed instead of 'int'. src/cAdt.cpp: small correction size_t -> unsigned long to avoid warnings at compilation.

+++ include/cManager.h
@@ -16,7 +16,7 @@ public:
   CManager();
   virtual ~CManager();
   int getGlobalIndex(std::string const& str_physics, int const& iProc, int const& iVertex);
-  void setGlobalIndexing(std::string str_physics, std::vector<std::vector<double> > index_range);
+  void setGlobalIndexing(std::string str_physics, std::vector<std::vector<int> > index_range);
   std::vector< std::vector< std::vector<int> > > globalIndexRange;
   int nPhyscis;
   int nIndex;
